flask_devops_app/requirements.txt
Flask>=2.2.3
pytest
pytest-flask
psycopg2-binary
redis
python-dotenv